pyproject.toml
[project]
name = "server-leads-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosmtplib>=5.0.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-postgres>=3.0.3",
    "openai>=2.15.0",
    "pgvector>=0.4.2",
    "psutil>=7.2.1",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.21",
    "torch>=2.0.0; platform_machine != 'x86_64' or sys_platform == 'darwin'",
    "pyjwt>=2.11.0",
    "bcrypt>=5.0.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
